+++ backend/db/postgres/modules/service-catalogue/module.sql
@@ -9,3 +9,4 @@
 \ir ../../schema/29_reduction_low_risk_cleanup.sql
 \ir ../../schema/30_reduction_domain_model_simplification.sql
 \ir ../../schema/32_final_reduction_sunset_cleanup.sql
+\ir ../../schema/36_service_sla_canonical.sql